Service File updates
Adjusted UI for status events
